# set vectorization factor for each IO tensor