--- a/csrc/scheduler/normalization_outer.cpp
+++ b/csrc/scheduler/normalization_outer.cpp
@@ -92,11 +92,13 @@ bool OuterPersistentKernelScheduler::canScheduleRunTime(
 
   auto reduced_tv = ir_utils::getSoleProducerTv(reduction_tvs.at(0));
 
-  const int64_t vectorization_factor = vectorize_helper::getVectorizationFactor(
-      runtime_info,
-      reduced_tv,
-      data_cache,
-      reduced_tv->nDims() - properties.inner_most_dimension_ndims).first;
+  const int64_t vectorization_factor =
+      vectorize_helper::getVectorizationFactor(
+          runtime_info,
+          reduced_tv,
+          data_cache,
+          reduced_tv->nDims() - properties.inner_most_dimension_ndims)
+          .first;
 
   // Minimum required multi reduction factor.
   const int64_t min_multi_reduction_factor = vectorization_factor *