--- a/csrc/scheduler/normalization_utils.h
+++ b/csrc/scheduler/normalization_utils.h
@@ -234,6 +234,8 @@ struct PersistentKernelProperties {
   PrimDataType index_type;
   bool has_exp_op;
   std::vector<TensorView*> persistent_buffers;
+  // map from TensorView to its max allowed vectorization factor
+  std::unordered_map<TensorView*, int64_t> vectorization_factor_map;
   std::string toString() const {
     std::stringstream ss;
     ss << "===== Persistent Kernel Properties ========\n"