… CMakeLists.txt
… benchmarks/python/global_params.py
… benchmarks/python/test_layer_norm_bwd_fp16_fp32_outputs.py
… benchmarks/python/test_layer_norm_fwd_fp16_fp32_outputs.py
… benchmarks/python/test_layer_norm_scale_shift_bwd.py
… csrc/scheduler/pointwise.cpp
… csrc/scheduler/reduction.cpp
… csrc/scheduler/reduction_heuristic.h
… csrc/scheduler/reduction_utils.cpp
… csrc/scheduler/reduction_utils.h
… csrc/scheduler/transpose.cpp
… csrc/scheduler/utils.cpp
… csrc/scheduler/utils.h
… csrc/scheduler/vectorize_helper.cpp
… csrc/scheduler/vectorize_helper.h
… tests/cpp/test_combined_inner_outer_reduction.cpp
… tests/cpp/test_gpu3.cpp
… tests/cpp/test_vectorization.cpp