@naoyam 2024-07-17
> last dim of T1 is further splitted & parallelized as S{8/4}, V{4}

Doesn't this mean each memory transaction will be strided? What does the performance look like?