@liqiangxl 2024-07-17
> > last dim of T1 is further splitted & parallelized as S{8/4}, V{4}
> 
> Doesn't this mean each memory transaction will be strided? What does the performance look like?

(1) Yes, it is strided, similar to dumping tmp results of fp32 to gmem in iter gouped grid reduction vectorized by 8 in  #2203 
(2) We don't have benchmarks designed for this case, working on this. Generally, it should depends on the fusion, e.g. if there is only 1 fp32 IO tensor and many other fp16 IO tensors, using a vectorization of 8 is probabally better than 4. However if there is only 1 fp16 but many fp32 IO tensors, it probabally better to use vectorization 4. It should be part of the heuristics.