@liqiangxl 2024-07-17
Checked performance in #2146:
```
(1) If enable this feature, inner outer normalization reduced from 325 us to 300 us
(2) If enable this feature, outer reduction kernel increased from 268 us to 340 us.
```
Seems depends on fusion and should leave it to the heuristics.
More benchmarks will be added later.